// apps/backend/src/generated/prisma/enums.ts
export const UserRole = {
  STUDENT: 'STUDENT',
  INSTRUCTOR: 'INSTRUCTOR',
  MANAGER: 'MANAGER',
  PRINCIPAL: 'PRINCIPAL',
  ADMIN: 'ADMIN',
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserStatus = {
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  ACTIVE: 'ACTIVE',
  REJECTED: 'REJECTED',
  INACTIVE: 'INACTIVE',
  DELETE_PENDING: 'DELETE_PENDING',
  DELETED: 'DELETED',
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER',
  UNDISCLOSED: 'UNDISCLOSED',
} as const;

export type Gender = (typeof Gender)[keyof typeof Gender];

export const SubjectMode = {
  THEORY: 'THEORY',
  PRACTICE: 'PRACTICE',
  MIXED: 'MIXED',
} as const;

export type SubjectMode = (typeof SubjectMode)[keyof typeof SubjectMode];

export const CourseStatus = {
  PLANNED: 'PLANNED',
  RECRUITING: 'RECRUITING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELED: 'CANCELED',
} as const;

export type CourseStatus = (typeof CourseStatus)[keyof typeof CourseStatus];